+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Coming Soon
+## v1.7.1
 
 ### fatcher